Factory: PR comments, adds new field in response for a liquidity pool, updates tests. Pair: adds new field for the factory response. Updates CHANGELOG.md
