+++ contracts/pair/src/storage.rs
@@ -105,6 +105,7 @@ pub struct PoolResponse {
 #[contracttype]
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct LiquidityPoolInfo {
+    pub pool_address: Address,
     pub pool_response: PoolResponse,
     pub total_fee_bps: i64,
 }
